pyproject.toml
[tool.poetry]
name = "boothit"
version = "0.1.0"
description = ""
authors = ["Tobias Brunner <tobias.brunner@vshn.ch>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.3"
requests = "^2.31.0"
python-dotenv = "^1.0.1"
bootstrap-flask = "^2.4.0"
flask-wtf = "^1.2.1"
email-validator = "^2.1.1"
brother-ql-web = "^0.2.1"


[tool.poetry.group.dev.dependencies]
black = "^24.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
